primera parte de la reestructuración
