+++ Server/bin/www.js
@@ -1,0 +1,9 @@
+#!/usr/bin/env node
+
+/**
+ *  Module dependencies
+ */
+
+
+// esto nos beneficia para poder importar de esta forma el objeto app dentro de los tests
+var app = require('../app');
