+++ Server/routes/index.js
@@ -1,0 +1,6 @@
+const express = require('express');
+const router = express.Router();
+const commerce = require('../apiServices/commerce/routes');
+
+
+router.use('/', commerce);
